# contracts/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "edu-ai-assistant-api",
    "description": "AI 教学助手后端服务。接口契约见 docs/api-contract.md；所有响应错误统一为 {error: {code, message, details, request_id}} 结构。",
    "version": "0.1.0"
  },
  "paths": {
    "/health/live": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "进程存活检查",
        "description": "进程可响应时返回 200，不访问数据库、对象存储或大模型。",
        "operationId": "liveness_health_live_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "500": {
            "description": "未预期的服务端错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/health/ready": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "依赖就绪检查",
        "description": "检查必需环境变量与数据库连通性。全部就绪返回 200，任一不可用返回 503 且响应体为统一错误结构。",
        "operationId": "readiness_health_ready_get",
        "responses": {
          "200": {
            "description": "配置齐备且数据库可达",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "500": {
            "description": "未预期的服务端错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "配置缺失或数据库不可达（SERVICE_UNAVAILABLE）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "注册教师或学生账户",
        "description": "第一版开放注册，无需邀请码或审批。并发注册由数据库唯一约束兜住，冲突时返回 409 AUTH_EMAIL_TAKEN。注册不返回令牌，注册完成后调用登录接口。",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "注册成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "500": {
            "description": "未预期的服务端错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "邮箱已被占用（AUTH_EMAIL_TAKEN）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "请求参数不合法（VALIDATION_ERROR）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "登录",
        "description": "邮箱登录，返回 Access Token 与 Refresh Token。登录失败按规范化邮箱限流；账号不存在与密码错误返回同一条错误信息。",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "500": {
            "description": "未预期的服务端错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "账号或密码错误（AUTH_INVALID_CREDENTIALS）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "请求参数不合法（VALIDATION_ERROR）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "失败次数过多，短时间内拒绝登录（AUTH_TOO_MANY_ATTEMPTS）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "刷新访问令牌",
        "description": "通过 JSON Body 提交 Refresh Token，不要求有效的 Access Token。只返回新的 Access Token，不轮换也不延长 Refresh Token 有效期。",
        "operationId": "refresh_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "刷新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshResponse"
                }
              }
            }
          },
          "500": {
            "description": "未预期的服务端错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "会话不存在、已过期或已撤销（AUTH_TOKEN_EXPIRED）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "请求参数不合法（VALIDATION_ERROR）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "注销当前会话",
        "description": "需要有效的 Access Token，并通过 JSON Body 指定要撤销的 Refresh Token 会话。只撤销该会话；已签发的 Access Token 仍可能有效至自身过期。",
        "operationId": "logout_api_v1_auth_logout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "注销成功，无响应体"
          },
          "500": {
            "description": "未预期的服务端错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Access Token 缺失、无效或已过期（AUTH_TOKEN_EXPIRED）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "会话不存在或不属于当前用户（RESOURCE_NOT_FOUND）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "请求参数不合法（VALIDATION_ERROR）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/users/me": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "当前用户资料",
        "operationId": "read_current_user_api_v1_users_me_get",
        "responses": {
          "200": {
            "description": "当前登录用户",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "500": {
            "description": "未预期的服务端错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Access Token 缺失、无效或已过期（AUTH_TOKEN_EXPIRED）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "ErrorBody": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorBody",
        "description": "错误体的内容部分。"
      },
      "ErrorCode": {
        "type": "string",
        "enum": [
          "AUTH_INVALID_CREDENTIALS",
          "AUTH_TOKEN_EXPIRED",
          "AUTH_EMAIL_TAKEN",
          "AUTH_TOO_MANY_ATTEMPTS",
          "ROLE_FORBIDDEN",
          "COURSE_FORBIDDEN",
          "RESOURCE_NOT_FOUND",
          "INVITE_CODE_INVALID",
          "UPLOAD_INVALID",
          "RUBRIC_SCORE_MISMATCH",
          "MATERIAL_NOT_READY",
          "ASSIGNMENT_NOT_OPEN",
          "GRADE_NOT_REVIEWED",
          "AI_JOB_FAILED",
          "VALIDATION_ERROR",
          "METHOD_NOT_ALLOWED",
          "INTERNAL_ERROR",
          "SERVICE_UNAVAILABLE"
        ],
        "title": "ErrorCode",
        "description": "机器可读的错误标识。"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "所有非 2xx 响应的统一结构。",
        "example": {
          "error": {
            "code": "COURSE_FORBIDDEN",
            "details": {},
            "message": "你没有访问该课程的权限",
            "request_id": "3f6a1c2e9d8b4a0f9e7c5b3a1d2f4e60"
          }
        }
      },
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "description": "登录请求（契约 2.1，JSON 提交）。"
      },
      "LoginResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          },
          "user": {
            "$ref": "#/components/schemas/UserSummary"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "refresh_token",
          "expires_in",
          "user"
        ],
        "title": "LoginResponse",
        "description": "登录响应（契约 2.1）。"
      },
      "LogoutRequest": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "minLength": 1,
            "title": "Refresh Token"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "LogoutRequest",
        "description": "注销请求（契约 2.5）：指定要撤销的当前会话。"
      },
      "RefreshRequest": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "minLength": 1,
            "title": "Refresh Token"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest",
        "description": "刷新请求（契约 2.4）：不要求有效的 Access Token。"
      },
      "RefreshResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "expires_in"
        ],
        "title": "RefreshResponse",
        "description": "刷新响应（契约 2.4）：只返回新的 Access Token。"
      },
      "RegisterRequest": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Password"
          },
          "display_name": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Display Name"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "email",
          "password",
          "display_name",
          "role"
        ],
        "title": "RegisterRequest",
        "description": "注册请求（契约 2.1）。"
      },
      "UserProfile": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "display_name",
          "role",
          "created_at"
        ],
        "title": "UserProfile",
        "description": "当前用户资料；用于 ``GET /users/me`` 与注册响应。"
      },
      "UserRole": {
        "type": "string",
        "enum": [
          "TEACHER",
          "STUDENT"
        ],
        "title": "UserRole",
        "description": "平台角色。\n\n平台角色只是默认权限，课程资源仍必须校验成员关系\n（见 ``docs/modules.md`` 第 1 节）。"
      },
      "UserSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          }
        },
        "type": "object",
        "required": [
          "id",
          "display_name",
          "role"
        ],
        "title": "UserSummary",
        "description": "登录响应中的用户概要，字段与契约 2.1 的示例一致。"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Bearer <access_token>",
        "scheme": "bearer"
      }
    }
  }
}